feat: add SOURCE_TRANSFORMER for source node handling; update editor transformers and example content
