+++ src/components/LexicalFocused/transformers/EditorTransformers.tsx
@@ -1,10 +1,8 @@
 import { TRANSFORMERS as BASE_TRANSFORMERS } from '@lexical/markdown';
 
 import { MARKDOWN_TRANSFORMERS } from './MarkdownTransformers';
-import { SOURCE_TRANSFORMER } from './SourceTransformer';
 
 export const EDITOR_TRANSFORMERS = [
   ...BASE_TRANSFORMERS,
   ...MARKDOWN_TRANSFORMERS,
-  SOURCE_TRANSFORMER,
 ];
