+++ src/playground/components/LexicalFocusedExample/LexicalFocusedExample.tsx
@@ -27,7 +27,7 @@ $$e=mc^2$$
 
 | Column 1 | Column 2 |
 | -------- | -------- |
-| Row 1    | Row 1    |
+| Row 1    | Row 1 [1]  |
 | Row 2    | Row 2    |
 
 `;
